Move normalizeTimeslot utility function to app-wide shared folder
